refactor: Extract all JSON request logic
