+++ src/request_body/mod.rs
@@ -1,5 +1,5 @@
 
 pub mod request_body;
 mod body;
-mod hurl_json_building;
+mod json_request_body;
